chore: use explicit dashboard data modules
