+++ app/dashboard/audit/page.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from "next-intl"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { AuditLogs } from "@/components/dashboard/audit-logs"
 import { useQuery } from "@tanstack/react-query"
-import { getAuditLogs } from "@/lib/dashboard-data"
+import { getAuditLogs } from "@/lib/dashboard-data.client"
 import { Loader2 } from "lucide-react"
 
 export default function DashboardAuditPage() {
