+++ app/dashboard/users/page.tsx
@@ -4,7 +4,7 @@ import { useTranslations } from "next-intl"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { Table, TableHeader, TableRow, TableHead, TableBody, TableCell } from "@/components/ui/table"
 import { useQuery } from "@tanstack/react-query"
-import { getUsers } from "@/lib/dashboard-data"
+import { getUsers } from "@/lib/dashboard-data.client"
 import { Loader2 } from "lucide-react"
 import { format } from "date-fns"
 import { Badge } from "@/components/ui/badge"
