+++ lib/dashboard-data.client.ts
@@ -6,6 +6,8 @@ import type {
   DashboardAnalytics,
   PendingReview,
   AdminAction,
+  Notification,
+  User,
 } from "./dashboard-types";
 
 /**
@@ -108,3 +110,96 @@ export async function getAdminActions(): Promise<
     };
   }
 }
+
+/**
+ * Fetches notifications for the current user
+ */
+export async function getNotifications(): Promise<
+  ServerActionResponse<Notification[]>
+> {
+  const supabase = createClient();
+
+  const {
+    data: { user }
+  } = await supabase.auth.getUser();
+
+  if (!user) {
+    return {
+      success: false,
+      message: "User not authenticated."
+    };
+  }
+
+  const { data, error } = await supabase
+    .from("notifications")
+    .select("id, message, created_at, read")
+    .eq("user_id", user.id)
+    .order("created_at", { ascending: false })
+    .limit(10);
+
+  if (error) {
+    return {
+      success: false,
+      message: `Failed to fetch notifications: ${error.message}`
+    };
+  }
+
+  return {
+    success: true,
+    message: "Notifications fetched successfully.",
+    data: data as Notification[]
+  };
+}
+
+/**
+ * Fetches list of users (admin view)
+ */
+export async function getUsers(): Promise<ServerActionResponse<User[]>> {
+  const supabase = createClient();
+
+  const { data, error } = await supabase
+    .from("users")
+    .select("id, email, role, created_at");
+
+  if (error) {
+    return {
+      success: false,
+      message: `Failed to fetch users: ${error.message}`
+    };
+  }
+
+  return {
+    success: true,
+    message: "Users fetched successfully.",
+    data: data as User[]
+  };
+}
+
+/**
+ * Fetches audit logs
+ */
+export async function getAuditLogs(
+  limit = 50
+): Promise<ServerActionResponse<AdminAction[]>> {
+  const supabase = createClient();
+
+  const { data, error } = await supabase
+    .from("audit_logs")
+    .select("*")
+    .order("created_at", { ascending: false })
+    .limit(limit);
+
+  if (error) {
+    return {
+      success: false,
+      message: `Failed to fetch audit logs: ${error.message}`
+    };
+  }
+
+  return {
+    success: true,
+    message: "Audit logs fetched successfully.",
+    data: data as AdminAction[]
+  };
+}
+
